+++ src/main/java/com/doublew2w/sbs/mybatis/executor/BaseExecutor.java
@@ -50,6 +50,14 @@ public abstract class BaseExecutor implements Executor {
   }
 
   @Override
+  public <E> List<E> query(
+      MappedStatement ms, Object parameter, RowBounds rowBounds, ResultHandler resultHandler)
+      throws SQLException {
+    BoundSql boundSql = ms.getBoundSql(parameter);
+    return query(ms, parameter, rowBounds, resultHandler, boundSql);
+  }
+
+  @Override
   public int update(MappedStatement ms, Object parameter) throws SQLException {
     log.info("executing an update");
     if (closed) {