+++ src/main/java/com/doublew2w/sbs/mybatis/type/DateTypeHandler.java
@@ -15,6 +15,15 @@ import java.util.Date;
  */
 public class DateTypeHandler extends BaseTypeHandler<Date> {
   @Override
+  protected Date getNullableResult(ResultSet rs, int columnIndex) throws SQLException {
+    Timestamp sqlTimestamp = rs.getTimestamp(columnIndex);
+    if (sqlTimestamp != null) {
+      return new Date(sqlTimestamp.getTime());
+    }
+    return null;
+  }
+
+  @Override
   protected Date getNullableResult(ResultSet rs, String columnName) throws SQLException {
     Timestamp sqlTimestamp = rs.getTimestamp(columnName);
     if (sqlTimestamp != null) {