+++ src/main/java/com/doublew2w/sbs/mybatis/type/LongTypeHandler.java
@@ -14,6 +14,11 @@ import java.sql.SQLException;
 public class LongTypeHandler extends BaseTypeHandler<Long> {
 
   @Override
+  protected Long getNullableResult(ResultSet rs, int columnIndex) throws SQLException {
+    return rs.getLong(columnIndex);
+  }
+
+  @Override
   protected Long getNullableResult(ResultSet rs, String columnName) throws SQLException{
     return rs.getLong(columnName);
   }