+++ src/test/java/com/doublew2w/sbs/mybatis/test/IActivityDaoApiTest.java
@@ -31,11 +31,29 @@ public class IActivityDaoApiTest {
   }
 
   @Test
-  public void test_queryActivityById() {
+  public void test_branch13_queryActivityById() {
     // 1. 获取映射器对象
     IActivityDao dao = sqlSession.getMapper(IActivityDao.class);
     // 2. 测试验证
     Activity res = dao.queryActivityById(100001L);
     log.info("测试结果：{}", JSON.toJSONString(res));
+  }
+
+  @Test
+  public void test_branch14_insert() {
+    // 1. 获取映射器对象
+    IActivityDao dao = sqlSession.getMapper(IActivityDao.class);
+
+    Activity activity = new Activity();
+    activity.setActivityId(10007L);
+    activity.setActivityName("测试活动");
+    activity.setActivityDesc("测试数据插入");
+    activity.setCreator("xiaofuge");
+
+    // 2. 测试验证
+    Integer res = dao.insert(activity);
+    sqlSession.commit();
+
+    log.info("测试结果：count：{} idx：{}", res, JSON.toJSONString(activity.getId()));
   }
 }